+++ eelbrain/plot/brain.py
@@ -1,5 +1,11 @@
 """Plot source space NDVars with mayavi/pysurfer."""
 # Author: Christian Brodbeck <christianbrodbeck@nyu.edu>
 
-from ._brain import (brain, cluster, dspm, surfer_brain, annot, bin_table,
-                     dspm_bin_table, copy, p_map, annot_legend, butterfly)
+from ._brain import (
+    SequencePlotter,
+    brain,
+    cluster, dspm, p_map, surfer_brain, annot, annot_legend,
+    butterfly,
+    bin_table, dspm_bin_table,
+    copy,
+)